ai-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
chromadb==0.5.5
google-genai==0.7.0
openai==1.43.0
tiktoken==0.7.0
httpx==0.27.0
pydantic==2.9.2